+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-02-14
+
+### Added
+
+- Add literal open and close settings for log messages
+
 ## [1.0.1] - 2025-02-14
 
 ### Fixed
@@ -19,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of codeLog+ extension
 
-[Unreleased]: https://github.com/ManuelGil/vscode-code-log-plus/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/ManuelGil/vscode-code-log-plus/compare/v1.1.0...HEAD
+[1.1.0]: https.//github.com/ManuelGil/vscode-code-log-plus/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/ManuelGil/vscode-code-log-plus/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ManuelGil/vscode-code-log-plus/releases/tag/v1.0.0